Disable the Rails/RakeEnvironment cop
rubocop-rails is included in this gem so that it can be pulled in to apps
that use the boxt_rubocop gem. But some of the Rails specific cops are not
applicable to this gem and can be disabled.

The `Rails/RakeEnvironment` for example will cause errors when we add
the :environment task dependency as the environment task is not present
in this gem